+++ CHANGES.md
@@ -37,7 +37,7 @@ Release Notes.
 * Fix RestTemplate plugin recording url tag with wrong port
 * Support collecting logs and forwarding through gRPC.
 * Support config `agent.sample_n_per_3_secs` can be changed in the runtime.
-* Support DNS period resolving mechanism to update backend service.
+* Support DNS periodic resolving mechanism to update backend service.
 
 #### OAP-Backend
 * Make meter receiver support MAL.